requirements.txt
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.4.0
chex==0.1.87
etils==1.10.0
flatbuffers==24.3.25
flax==0.10.0
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.1
h5py==3.12.1
humanize==4.11.0
idna==3.10
importlib_resources==6.4.5
jax==0.4.34
jaxlib==0.4.34
keras==3.6.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
msgpack==1.1.0
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
opt_einsum==3.4.0
optax==0.2.3
optree==0.13.0
orbax-checkpoint==0.8.0
packaging==23.2
pandas==2.2.3
pillow==11.0.0
protobuf==4.25.5
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.3
scipy==1.14.1
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow_decision_forests==1.9.0
tensorflowjs==4.22.0
tensorstore==0.1.67
termcolor==2.5.0
tf_keras==2.16.0
toolz==1.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.16.0
wurlitzer==3.1.1
zipp==3.20.2
